Move logic to another file
